fix css and move handleSendMessage from Chat Component to InputBox Component
