+++ src/components/Messages.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import ScrollToBottom from 'react-scroll-to-bottom';
 import Message from './Message';
