--- src/styles/_app.css
@@ -1,1 +1,0 @@
-
