+++ src/styles/application.css
@@ -1,4 +1,3 @@
-@import '_App';
 @import '_Home';
 @import '_MainNav';
 @import '_Chat';
